[COMMIT]: Nome do component renomeado para AzReorderable.
