+++ package.json
@@ -18,7 +18,7 @@
   "author": "Charles Viegas",
   "license": "MIT",
   "dependencies": {
-    "@azinformatica/pdfjs-dist": "~2.10.377",
+    "@azinformatica/pdfjs-dist": "~2.10.380",
     "accounting": "~0.4.1",
     "all-contributors-cli": "~6.20.0",
     "axios": "~0.27.2",
@@ -69,7 +69,7 @@
     "webpack-dev-server": "~4.9.1"
   },
   "peerDependencies": {
-    "@azinformatica/pdfjs-dist": "~2.10.377",
+    "@azinformatica/pdfjs-dist": "~2.10.380",
     "axios": "~0.27.2",
     "moment-timezone": "~0.5.34",
     "scrollmonitor": "1.2.4",
